BUG: cast test data to int, avoid patsy long int name

+++ statsmodels/stats/tests/test_anova.py
@@ -72,7 +72,7 @@ kidney_table = StringIO("""Days      Duration Weight ID
 """)
 
 kidney_table.seek(0)
-kidney_table = read_table(kidney_table, sep="\s+")
+kidney_table = read_table(kidney_table, sep="\s+").astype(int)
 
 class TestAnovaLM(object):
     @classmethod
